# docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
...
  dto.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.SignUpRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@normark.com
    name: API Support
  description: A comprehensive trading journal API for tracking and analyzing trading
    performance
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Normark Trading Journal API
  version: "1.0"
paths:
  /api/v1/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated with access and refresh tokens
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/sign-up:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, username and password
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered user with access and refresh tokens
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"